# apps/studio/frontend/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Junjo AI Studio",
    "description": "The Junjo AI Studio API: observability, execution evidence, and evaluation control.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.86.0"
  },
  "paths": {
...
    "/api/v1/cli-sign-ins": {
      "post": {
        "tags": [
          "cli-sign-ins"
        ],
        "summary": "Start a CLI sign-in.",
        "description": "Public: the terminal that asks has no credential yet. Starting also\ndeletes every expired sign-in.",
        "operationId": "start_cli_sign_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CliSignInStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The started sign-in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliSignInStarted"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cli-sign-ins/token": {
      "post": {
        "tags": [
          "cli-sign-ins"
        ],
        "summary": "Collect the token of a CLI sign-in.",
        "description": "Public: the device code is the credential. The token is returned once,\nand the sign-in is deleted with it. Until then the answer is a 400 whose\ncode says why: `authorization_pending` while the sign-in is undecided,\n`access_denied` once when it was denied, and `expired_token` when the\ndevice code is unknown, already used, or expired.",
        "operationId": "collect_cli_sign_in_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CliSignInTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The token. It is returned once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliSignInToken"
                }
              }
            }
          },
          "400": {
            "description": "Nothing was collected: `authorization_pending`, `access_denied`, or `expired_token`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cli-sign-ins/{user_code}": {
      "get": {
        "tags": [
          "cli-sign-ins"
        ],
        "summary": "Get a pending CLI sign-in, for the approval page.",
        "operationId": "get_cli_sign_in",
        "parameters": [
          {
            "name": "user_code",
            "in": "path",
            "description": "The user code, in either case, with or without its hyphen.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 9,
              "minLength": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pending sign-in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliSignInRead"
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "CLI sign-in not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cli-sign-ins/{user_code}/approve": {
      "post": {
        "tags": [
          "cli-sign-ins"
        ],
        "summary": "Approve a pending CLI sign-in.",
        "description": "This mints a developer access token with the name and the scopes the\nterminal asked for and no expiry. It belongs to the approving user, and\nthe terminal collects it.",
        "operationId": "approve_cli_sign_in",
        "parameters": [
          {
            "name": "user_code",
            "in": "path",
            "description": "The user code, in either case, with or without its hyphen.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 9,
              "minLength": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The minted token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliSignInApproved"
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "CLI sign-in not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cli-sign-ins/{user_code}/deny": {
      "post": {
        "tags": [
          "cli-sign-ins"
        ],
        "summary": "Deny a pending CLI sign-in. The terminal is told once and gets no token.",
        "operationId": "deny_cli_sign_in",
        "parameters": [
          {
            "name": "user_code",
            "in": "path",
            "description": "The user code, in either case, with or without its hyphen.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 9,
              "minLength": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The sign-in was denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "CLI sign-in not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sign-in": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with an email address and a password.",
        "operationId": "sign_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The user is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "The credentials are not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CliSignInApproved": {
        "type": "object",
        "description": "The developer access token that approving a CLI sign-in minted.",
        "required": [
          "token_id",
          "token_name"
        ],
        "properties": {
          "token_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "token_name": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "CliSignInRead": {
        "type": "object",
        "description": "A pending CLI sign-in, as the approval page shows it.",
        "required": [
          "user_code",
          "client_name",
          "scopes",
          "expires_at"
        ],
        "properties": {
          "user_code": {
            "type": "string",
            "description": "The code the terminal shows. A person compares the two before\napproving.",
            "examples": [
              "WDJB-MJHT"
            ],
            "maxLength": 9,
            "minLength": 9
          },
          "client_name": {
            "type": "string",
            "description": "What the terminal calls itself. The terminal chose this text and\nStudio has not verified it.",
            "maxLength": 256,
            "minLength": 1
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationTokenScope"
            },
            "description": "The scopes the token will hold."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the sign-in can no longer be approved."
          }
        },
        "additionalProperties": false
      },
      "CliSignInStart": {
        "type": "object",
        "description": "The request to start a CLI sign-in.",
        "required": [
          "client_name",
          "scopes"
        ],
        "properties": {
          "client_name": {
            "type": "string",
            "description": "What the terminal calls itself. The approval page shows it, and it\nbecomes the name of the token.",
            "examples": [
              "junjo CLI on build-host"
            ],
            "maxLength": 256,
            "minLength": 1
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationTokenScope"
            },
            "description": "The scopes the token will hold.",
            "maxItems": 3,
            "minItems": 1
          }
        },
        "additionalProperties": false
      },
      "CliSignInStarted": {
        "type": "object",
        "description": "A started CLI sign-in.",
        "required": [
          "device_code",
          "user_code",
          "verification_path",
          "expires_in",
          "interval"
        ],
        "properties": {
          "device_code": {
            "type": "string",
            "description": "The secret the CLI keeps. It collects the token.",
            "examples": [
              "jdev_0123456789_abcdefghijklmnopqrstuvwxyz-ABCDEFGHIJKLMNOPQRSTUVWXYZ"
            ]
          },
          "user_code": {
            "type": "string",
            "description": "The code the CLI shows. A person confirms it in the browser.",
            "examples": [
              "WDJB-MJHT"
            ],
            "maxLength": 9,
            "minLength": 9
          },
          "verification_path": {
            "type": "string",
            "description": "The approval page. The CLI appends it to its configured Studio origin\nand adds `?code=<user_code>`.",
            "examples": [
              "/cli-sign-in"
            ]
          },
          "expires_in": {
            "type": "integer",
            "description": "How many seconds the sign-in lives.",
            "examples": [
              900
            ],
            "minimum": 1
          },
          "interval": {
            "type": "integer",
            "description": "How many seconds the CLI waits between polls.",
            "examples": [
              5
            ],
            "minimum": 1
          }
        },
        "additionalProperties": false
      },
      "CliSignInToken": {
        "type": "object",
        "description": "The developer access token an approved CLI sign-in minted. It is returned\nonce.",
        "required": [
          "access_token",
          "token_type",
          "token_id",
          "scopes",
          "expires_at"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The developer access token value.",
            "examples": [
              "jcli_0123456789_abcdefghijklmnopqrstuvwxyz-ABCDEFGHIJKLMNOPQRSTUVWXYZ"
            ]
          },
          "token_type": {
            "type": "string",
            "description": "How a collected token is presented: always as a bearer credential.",
            "const": "bearer"
          },
          "token_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationTokenScope"
            }
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the token stops working. Null: a token minted by a CLI sign-in\nworks until it is deleted."
          }
        },
        "additionalProperties": false
      },
      "CliSignInTokenRequest": {
        "type": "object",
        "description": "The request to collect the token of a CLI sign-in.",
        "required": [
          "device_code"
        ],
        "properties": {
          "device_code": {
            "type": "string",
            "description": "The device code the sign-in was started with.",
            "pattern": "^jdev_[A-Za-z0-9_-]{64}$"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable machine-readable code.",
            "examples": [
              "not_found"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable explanation."
          }
        }
      },
...
      "EvaluationTokenScope": {
        "type": "string",
        "description": "Authorities supported by the evaluation control/query credential.",
        "enum": [
          "evaluation:read",
          "evaluation:write",
          "evidence:read"
        ]
      },
...
      "SignInRequest": {
        "type": "object",
        "description": "The request to sign in.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "description": "The response after a successful user operation.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "EvaluationControlToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A separately scoped Studio evaluation-control token. Studio ingestion API keys are not accepted."
      }
    }
  }
}